refactor: place citis fetach calls into api folder
